requirements.txt
Flask==2.3.3
gunicorn==21.2.0
Werkzeug==3.1.4
Jinja2==3.1.6
click==8.1.7
itsdangerous==2.2.0
blinker==1.9.0
packaging==25.0
MarkupSafe==3.0.3
Flask-Login==0.6.3
requests==2.31.0
requests-oauthlib==1.3.1
python-dotenv==1.0.0
pytest==8.3.3
Authlib==1.2.0
